Initial ebuild for geoip2lookup extension
This ebuild is heavily taken from the current one for pgtelemetry.

Note:  the XS dependency is technically optional.  In many cases I might be
persuaded to put it in a USE flag is there is an absolute consensus that this
is the way to go.  However the counter-argument is that the C implementation
is so much faster than the pure Perl one that putting the pure Perl one in
the database is just a bad idea.  So for this reason I am making both of these
dependencies required.