Add postgresql-10.4-r71 ebuild with -dbg SLOT suffix for cassert
Revision is picked specifically to be smaller than 101 to prevent any
potential problems during dependency resolution.

Admittedly, suffixing the SLOT is not the optimal way to do this. In the
future we could use virtuals for more fine-grained install control. See
the relevant issue for more context. Ref: adjust/infrastructure#3345